did some testing of simple math functions. min/max implemented with < and > do not handle overflows of floating point within the argument which might lead to problems. therefore:
IUMIN / IUMAX now uses < > checks if called on CPU data and
min(), max() on GPU stuff as those should use the correct types.

sqr is now in the namespace iu:: for safety....





git-svn-id: https://svn.icg.tugraz.at/svn/vmgpu/trunk/vmlibraries/imageutilities@11550 358b74ad-2626-0410-af80-f4c3c349ae55

+++ src/iuinteraction/draw.cu
@@ -81,7 +81,7 @@ IuStatus CUDAdrawLine(iu::ImageGpu_8u_C1 *image, int x_start, int y_start,
                       int x_end, int y_end, int line_width, unsigned char value)
 {
   // adapt brush size_x
-  float brush_size = line_width*line_width;
+  int brush_size = line_width*line_width;
 
   // extract some variables
   int width = image->width();
@@ -104,7 +104,7 @@ IuStatus CUDAdrawLine(iu::ImageGpu_8u_C1 *image, int x_start, int y_start,
   dim3 dimGrid(iu::divUp(size_x, dimBlock.x), iu::divUp(size_y, dimBlock.y));
 
   // Call drawing algorithm
-  drawLineKernel<<<dimGrid, dimBlock>>>(image->data(), 
+  drawLineKernel<<<dimGrid, dimBlock>>>(image->data(),
                                         x_start, y_start, x_end, y_end,
                                         min_x, min_y, width, height, image->stride(),
                                         brush_size, value);
